call printUsage() instead of printUsageSummary(). this was my plan all along, i just needed to figure out where to get a CmdLineParser object to pass. i think what i'm doing to create it is okay, and it seems to work, but it was somewhat of a guess.
one slight difference between this case (calling --help explicitly) and the
implicit usage (e.g. when you get the command wrong) is that in this case
we're only printing out the options for this command, but in the other case
the general options to the jenkins CLI are also output.  i can't decide
whether i'd want those or not, but i don't know how to control them anyway, so
it's fine how it is now.