Add test to verify JCasC compatibility
For a guideline on JCasC see:
https://github.com/jenkinsci/configuration-as-code-plugin/blob/configuration-as-code-1.2/docs/PLUGINS.md

JCasC requires Jenkins 2.60.3. Therefore upgrading the minimum Jenkins version.
Jenkins 2.x uses Groovy 2.4.x. Therefore upgrading the Spock dependency.

Excluding the 'groovy-all' module as Spock requires 2.4.15 but Jenkins requires 2.4.8.
That's a known issue with Spock - see https://github.com/spockframework/spock/issues/723.
This results in both modules on the classpath when tests are run in IntelliJ IDEA.

Actually if this plugin would be used without JCasC, the Jenkins version
wouldn't require an upgrade. But for testing it's required.