Avoid applying CCUD plugin in Gradle tests
The CCUD plugin will read a `DEVELOCITY_URL` env var and override all other settings.
This is problematic on CI, where the TeamCity plugin sets this value for DV injection.
We should only apply CCUD in tests where it is actually required.

This issue will be addressed in a future TeamCity plugin, where the the DV injection
env vars have been changed so as not to intersect with CCUD gradle plugin overrides.
Applying CCUD only when necessary still makes sense with this fix.